ferris-wheel compile-fail

+++ ferris-wheel/compile-fail/move.rs
@@ -1,6 +1,6 @@
 // FIXME: Prevent this file from compiling! Diff budget: 1 line.
 #[derive(Clone, Copy)]
-struct MyType(usize);
+struct MyType(f64);
 
 // Note: do not modify this function.
 fn main() {
